+++ src/main/kotlin/dev/bombardy/bardybot/audio/LoadResultHandler.kt
@@ -1,13 +1,12 @@
 package dev.bombardy.bardybot.audio
 
 import com.sedmelluq.discord.lavaplayer.player.AudioLoadResultHandler
+import com.sedmelluq.discord.lavaplayer.source.AudioSourceManager
 import com.sedmelluq.discord.lavaplayer.tools.FriendlyException
 import com.sedmelluq.discord.lavaplayer.track.AudioPlaylist
 import com.sedmelluq.discord.lavaplayer.track.AudioTrack
-import com.sedmelluq.discord.lavaplayer.source.AudioSourceManager
 import dev.bombardy.bardybot.services.TrackService
 import net.dv8tion.jda.api.entities.Member
-import net.dv8tion.jda.api.entities.MessageChannel
 import net.dv8tion.jda.api.entities.TextChannel
 
 /**