+++ src/main/kotlin/dev/bombardy/bardybot/config/BotConfig.kt
@@ -2,7 +2,6 @@ package dev.bombardy.bardybot.config
 
 import org.springframework.boot.context.properties.ConfigurationProperties
 import org.springframework.boot.context.properties.ConstructorBinding
-import org.springframework.context.annotation.Configuration
 
 /**
  * Wrapper class for configuration values stored in application.yml